refactor: move FilePickerService inside Eppie.App.UI.Behaviors
References: Eppie-io/Eppie-App#382
